KTB-07: Completed tasks for stage 07
