+++ src/main/kotlin/Word.kt
@@ -1,0 +1,7 @@
+package org.example
+
+data class Word(
+    val englishText: String,
+    val russianText: String,
+    var correctAnswersCount: Int = 0,
+)
