+++ src/main/kotlin/showStatistics.kt
@@ -1,0 +1,14 @@
+package org.example
+
+fun MutableList<Word>.showStatistics() {
+
+    val mutableList = this.filter {
+        it.correctAnswersCount >= MAX_CORRECT_ANSWER_COUNT
+    }
+
+    val allWords = this.size
+    val count = mutableList.size
+    val percent = ((count.toDouble() / allWords) * 100).toInt()
+
+    println("Выучено $count из $allWords слов | $percent%")
+}
